Made solution mobile friendly. Caveats below:
Unfortunately this solution doesn't work when resizing from larger than
mobile threshold to small than it (an esoteric case). It does work from
smaller to larger and when the page is loaded and not resized fine,
however.

This needs a fix. It will probably mean switching from dropotron to
something else since dropotron doesn't allow for deactivating or
unloading itself after initialization.